pyproject.toml
[project]
name = "fish-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.1.0",
    "fairseq>=0.12.2",
    "gtts>=2.5.4",
    "ipykernel>=6.29.5",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.4",
    "langchain-groq>=0.2.1",
    "langchain-ollama>=0.2.0",
    "langchain-together>=0.2.0",
    "langgraph>=0.2.50",
    "load-dotenv>=0.1.0",
    "nltk>=3.9.1",
    "numpy==2.0",
    "openai-whisper>=20240930",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "pyttsx3>=2.98",
    "rich>=13.9.4",
    "scipy>=1.14.1",
    "sentencepiece>=0.2.0",
    "simpleaudio>=1.0.4",
    "sounddevice>=0.5.1",
    "soundfile>=0.12.1",
    "torch>=2.5.1",
    "transformers>=4.46.2",
]
